requirements.txt
alembic==1.9.2
bcrypt==4.0.1
fastapi==0.100.0
jwt==1.3.1
SQLAlchemy==1.4.41
requests==2.28.2
passlib==1.7.4
XlsxWriter==3.0.3
uvicorn==0.15.0
pydantic-settings==2.5.2
fastapi_pagination==0.12.27
pytz==2024.2
jose=1.0.0
python-jose==3.3.0
pymysql==1.1.1
email_validator==2.2.0
python-multipart==0.0.10

